Refactor: Extract codegen mapping logic from domain models
- Created `src/codegen/mappings.rs` to handle all Domain->Rust conversions
- Removed `to_rust()` methods from `GlossaType`, `Case`, `BinaryOp`, `UnaryOp`
- Moved name sanitization and transliteration to `mappings.rs`
- Updated integration tests to use the new mapping API
- Decoupled `semantic` and `morphology` modules from Rust code generation details
